Extract most of the readme check into a separate validator class
